Balacing first pass alex

+++ Assets/Dossiers Personnels/Sarah/Script/S_O/CrossbowStats_SO.cs
@@ -12,5 +12,5 @@ public class CrossbowStats_SO : ScriptableObject
     [field: SerializeField]
     public float kockbackForce { get; private set; }
     [field: SerializeField]
-    public float cost { get; private set; }
+    public int cost { get; private set; }
 }
