+++ Assets/Dossiers Personnels/Sarah/Script/S_O/MonsterStats_SO.cs
@@ -17,6 +17,6 @@ public abstract class MonsterStats_SO : ScriptableObject
     [field: SerializeField]
     public float kockbackForce { get; private set; }
     [field: SerializeField]
-    public float cost { get; private set; }
+    public int cost { get; private set; }
 
 }
